Update DdevSettingUpTypo3.rst
If you dont use ddev stop and just restart docker or kill docker it might corrupt the database.

This happend fluently for me, because if i work with docker beside projects setup with ddev, i dont pay that much attention if there is still a ddev project running and just restart. 
On windows there is still alot of pain with the docker nfs stuff, so restarting docker is something windows users do pretty often.  And as described in the linked ddev github issues is might be relevant for mac too.